Move DatabaseTestCase to testtools
